+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 ### Added
 - [55](https://github.com/picimako/just-kitting/issues/55): Added `CheckLicense.java` to the list of diffable resources.
 
+### Fixed
+- Now the value of the `gradleVersion` property is queried right when clicking the Update Gradle version line marker,
+and not saved in the line marker at creation.
+
 ## [1.1.1]
 ### Fixed
 - Fixed the K2 compiler compatibility configuration.